feat(desktop): add Calibration tab shell
